feat(customer-portal): prompt for a card before resuming
