+++ clients/apps/web/src/components/CustomerPortal/UncancelSubscriptionButton.tsx
@@ -1,15 +1,10 @@
 'use client'
 
 import { toast } from '@/components/Toast/use-toast'
-import {
-  useCustomerPaymentMethods,
-  useCustomerUncancelSubscription,
-} from '@/hooks/queries/customerPortal'
-import { PolarEmbedPaymentMethod } from '@polar-sh/checkout/payment-method'
+import { useCustomerUncancelSubscription } from '@/hooks/queries/customerPortal'
+import { useRequirePaymentMethod } from '@/hooks/useRequirePaymentMethod'
 import type { Client, schemas } from '@polar-sh/client'
 import { Button } from '@polar-sh/orbit'
-import { useQueryClient } from '@tanstack/react-query'
-import { useTheme } from 'next-themes'
 
 const UncancelSubscriptionButton = ({
   api,
@@ -22,11 +17,9 @@ const UncancelSubscriptionButton = ({
   customerSessionToken: string
   onUncancelled: () => Promise<void>
 }) => {
-  const theme = useTheme()
-  const queryClient = useQueryClient()
   const uncancelSubscription = useCustomerUncancelSubscription(api)
-  const { data: paymentMethods, isPending: paymentMethodsPending } =
-    useCustomerPaymentMethods(api)
+  const { withPaymentMethod, isPending: paymentMethodsPending } =
+    useRequirePaymentMethod(api, customerSessionToken)
 
   const uncancel = async () => {
     const { error } = await uncancelSubscription.mutateAsync({
@@ -46,27 +39,9 @@ const UncancelSubscriptionButton = ({
     await onUncancelled()
   }
 
-  const onClick = async () => {
-    if (paymentMethods && paymentMethods.items.length > 0) {
-      await uncancel()
-      return
-    }
-
-    const embed = await PolarEmbedPaymentMethod.create({
-      sessionToken: customerSessionToken,
-      theme: theme.resolvedTheme === 'dark' ? 'dark' : 'light',
-    })
-    embed.addEventListener('success', async () => {
-      await queryClient.invalidateQueries({
-        queryKey: ['customer_payment_methods'],
-      })
-      await uncancel()
-    })
-  }
-
   return (
     <Button
-      onClick={onClick}
+      onClick={() => withPaymentMethod(uncancel)}
       loading={uncancelSubscription.isPending || paymentMethodsPending}
       disabled={uncancelSubscription.isPending || paymentMethodsPending}
     >
